wip: Fix typescript issues
